requirements.txt
Flask==3.1.3
Flask-SocketIO==5.6.1
Flask-SQLAlchemy==3.1.1
Flask-Cors==6.0.2
python-socketio==5.16.1
python-engineio==4.13.1
Werkzeug==3.1.8
SQLAlchemy==2.0.49
python-bidi==0.4.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0

# Testes
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
